requirements.txt
torch>=1.12
torchvision
numpy
pandas
scikit-learn
matplotlib
Pillow
tqdm
streamlit
seaborn
datasets
